refactor(kvp): rename KvpRepository → KvpService
KvpRepository never extended BaseRepository — it was a thin wrapper
around Spatie\Valuestore placed in app/Repositories/ by convention.
Renamed to KvpService and moved to app/Services/ so the repositories
namespace empties out cleanly.

Class body unchanged. All 7 caller files (helpers.php, 3 services,
Filament Maintenance page, 2 tests) updated to the new FQCN.

Part of Phase 1b of the Prettus repository removal.
