+++ app/Services/CronService.php
@@ -3,7 +3,6 @@
 namespace App\Services;
 
 use App\Contracts\Service;
-use App\Repositories\KvpRepository;
 use DateTime;
 use DateTimeZone;
 use Illuminate\Support\Facades\Log;
@@ -12,7 +11,7 @@ use Symfony\Component\Process\PhpExecutableFinder;
 class CronService extends Service
 {
     public function __construct(
-        private readonly KvpRepository $kvpRepo
+        private readonly KvpService $kvpRepo
     ) {}
 
     /**
